melhorando a validação do back-end

+++ api/src/app/controllers/ContactController.js
@@ -1,4 +1,5 @@
 const ContactsRepository = require('../repositories/ContactsRepository');
+const isValidUUID = require('../utils/isValidUUID')
 
 class ContactController {
   async index(request, response) {
@@ -10,6 +11,11 @@ class ContactController {
 
   async show(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id' })
+    }
+
     const contact = await ContactsRepository.findById(id);
 
     if (!contact) {
@@ -28,14 +34,23 @@ class ContactController {
       return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactsExits = await ContactsRepository.findByEmail(email);
-
-    if (contactsExits) {
-      return response.status(400).json({ error: 'This email is alredy in use' });
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category' })
     }
 
+    if(email) {
+      const contactsExits = await ContactsRepository.findByEmail(email);
+      if (contactsExits) {
+        return response.status(400).json({ error: 'This email is alredy in use' });
+      }
+    }
+
+
     const contact = await ContactsRepository.create({
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
 
     response.status(201).json(contact);
@@ -47,29 +62,46 @@ class ContactController {
       name, email, phone, category_id,
     } = request.body;
 
-    const contactExits = await ContactsRepository.findById(id);
-    if (!contactExits) {
-      return response.status(404).json({ error: 'Contact not found' });
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id' })
+    }
+
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category' })
     }
 
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactByEmail = await ContactsRepository.findByEmail(email);
-    if (contactByEmail && contactByEmail.id !== id) {
-      return response.status(400).json({ error: 'This email is alredy in use' });
+    const contactExits = await ContactsRepository.findById(id);
+    if (!contactExits) {
+      return response.status(404).json({ error: 'Contact not found' });
+    }
+
+    if(email) {
+      const contactByEmail = await ContactsRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id !== id) {
+        return response.status(400).json({ error: 'This email is alredy in use' });
+      }
     }
 
     const contact = await ContactsRepository.update(id, {
-      name, email, phone, category_id,
-    });
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
+     });
 
     response.json(contact);
   }
 
   async delete(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id' })
+    }
 
     await ContactsRepository.delete(id);
     response.sendStatus(204);
